[toolbox-sca-120] Refactor Script Parser
